+++ src/prelude.rs
@@ -1,2 +1,4 @@
+#[cfg(feature = "async")]
+pub use crate::AsyncWriteOnlyDataCommand as _display_interface_AsyncWriteOnlyDataCommand;
 pub use crate::DisplayError as _display_interface_DisplayError;
 pub use crate::WriteOnlyDataCommand as _display_interface_WriteOnlyDataCommand;